Splitting conftest.py into several smaller files.
